+++ search_providers/__init__.py
@@ -5,11 +5,15 @@ from search_providers.base_search_provider import (
     response_to_dict,
 )
 from search_providers.factory import build_search_provider
+from search_providers.bing_search_provider import BingSearchProvider
+from search_providers.duckduckgo_search_provider import DuckDuckGoSearchProvider
 from search_providers.fixture_search_provider import FixtureSearchProvider
 from search_providers.http_search_provider import HttpSearchProvider
 
 __all__ = [
     "BaseSearchProvider",
+    "BingSearchProvider",
+    "DuckDuckGoSearchProvider",
     "FixtureSearchProvider",
     "HttpSearchProvider",
     "SearchResponse",